requirements.txt
Django>=1.2.3
-e git://github.com/django-extensions/django-extensions.git#egg=django-extensions
django_debug_toolbar
django-staticfiles
django-sentry
django-taggit
pisa
